+++ babel.config.js
@@ -4,6 +4,6 @@ module.exports = {
       targets: {
         node: 'current'
       }
-    }]
+    } ]
   ]
 };