+++ packages/cli/__fixtures__/EmptyCommand.js
@@ -1,0 +1,3 @@
+import Command from '../lib/commands/command';
+
+export default class EmptyCommand extends Command {}