+++ packages/cli/__mocks__/@averjs/core.js
@@ -1,0 +1,11 @@
+export const mockRun = jest.fn();
+export const mockBuild = jest.fn();
+
+const core = jest.fn().mockImplementation(() => {
+  return {
+    run: mockRun,
+    build: mockBuild
+  };
+});
+
+export default core;