+++ packages/config/__tests__/index.spec.js
@@ -1,5 +1,5 @@
 import path from 'path';
-import { getAverjsConfig } from '../index';
+import { getAverjsConfig } from '../lib/index';
 
 beforeEach(() => {
   process.env.PROJECT_PATH = path.resolve(__dirname, '../__fixtures__/src');
@@ -8,15 +8,11 @@ beforeEach(() => {
 test('should change progressbar option to false', () => {
   const config = getAverjsConfig();
   expect(config.progressbar).toBe(false);
-});
-
-test('should change nested attribute extract to true', () => {
-  const config = getAverjsConfig();
   expect(config.webpack.css.extract).toBe(true);
 });
 
-test('should return a default config when no config file is present', () => {
-  process.env.PROJECT_PATH = '';
+test('should match default snapshot config when no config file is present', () => {
+  process.env.PROJECT_PATH = '/aver';
   const config = getAverjsConfig();
-  expect(config.progressbar).toBe(true);
+  expect(config).toMatchSnapshot();
 });