+++ packages/core/app/index.js
@@ -1,3 +1,0 @@
-const Core = require('@averjs/core');
-const core = new Core();
-core.run();