+++ packages/init/lib/app/src/pages/Home/components/Home.vue
@@ -1,0 +1,29 @@
+<template>
+    <div>
+        <div :class="$style.test">Hallo</div>
+        <logo></logo>
+    </div>
+</template>
+
+<script>
+    import Vue              from 'vue';
+    import { Component }    from 'vue-property-decorator';
+    import Logo             from '@/components/Logo';
+	
+	@Component({
+        components: {
+            Logo
+        }
+    })
+    export default class Home extends Vue {
+        mounted() {
+            console.log('test');
+        }
+    }
+</script>
+
+<style lang="scss" module>
+    .test {
+        color: yellowgreen
+    }
+</style>