+++ packages/vue-app/lib/entry-server.js
@@ -1,6 +1,7 @@
+import { createApp } from './app';
 import Vue from 'vue';
 import App from '@/App.vue';
-import { createApp } from './app';
+import { composeComponentOptions } from './utils';
 
 Vue.prototype.$auth = null;
 Vue.prototype.$modernizr = {};
@@ -15,6 +16,7 @@ export default async context => {
     const meta = app.$meta();
 
     router.push(url);
+    context.meta = meta;
 
     await new Promise((resolve, reject) => router.onReady(resolve, reject));
     const matchedComponents = router.getMatchedComponents();
@@ -24,8 +26,6 @@ export default async context => {
       error.code = 404;
       throw error;
     }
-
-    context.meta = { inject: function() { Object.assign(this, meta.inject()); } };
         
     for (const r of mixinContext.keys()) {
       const EntryServerMixin = mixinContext(r).default;
@@ -40,24 +40,24 @@ export default async context => {
         await store.dispatch(key, context);
       }
     }
-        
-    for (const { options: { asyncData, props } } of matchedComponents) {
+
+    for (const component of matchedComponents) {
+      const { asyncData } = composeComponentOptions(component);
+
       if (typeof asyncData === 'function' && asyncData) {
         await asyncData({
           store,
           route: router.currentRoute,
-          data: props,
           isServer: true
         });
       }
     }
 
-    const { options: { asyncData, props } } = App;
+    const { asyncData } = composeComponentOptions(App);
     if (typeof asyncData === 'function' && asyncData) {
       await asyncData({
         store,
         route: router.currentRoute,
-        data: props,
         isServer: true
       });
     }