+++ packages/shared-utils/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-4](https://github.com/exreplay/aver.js/compare/v2.0.0-3...v2.0.0-4) (2019-11-23)
+
+**Note:** Version bump only for package @averjs/shared-utils
+
+
+
+
+
+# [2.0.0-3](https://github.com/exreplay/aver.js/compare/v2.0.0-2...v2.0.0-3) (2019-09-19)
+
+**Note:** Version bump only for package @averjs/shared-utils
+
+
+
+
+
+# [2.0.0-2](https://github.com/exreplay/aver.js/compare/v2.0.0-1...v2.0.0-2) (2019-09-18)
+
+**Note:** Version bump only for package @averjs/shared-utils
+
+
+
+
+
+# [2.0.0-1](https://github.com/exreplay/aver.js/compare/v2.0.0-0...v2.0.0-1) (2019-09-03)
+
+**Note:** Version bump only for package @averjs/shared-utils
+
+
+
+
+
+# [2.0.0-0](https://github.com/exreplay/aver.js/compare/v1.5.3...v2.0.0-0) (2019-09-03)
+
+
+### Features
+
+* add rollup to build packages ([#56](https://github.com/exreplay/aver.js/issues/56)) ([537faa6](https://github.com/exreplay/aver.js/commit/537faa6))
+
+
+
+
+
 ## [1.5.3](https://github.com/exreplay/aver.js/compare/v1.5.2...v1.5.3) (2019-06-13)
 
 **Note:** Version bump only for package @averjs/shared-utils